Prevent nested directory symlinks in create_dir_symlinks().
Sponsored by:	The FreeBSD Foundation

git-svn-id: file:///Users/gjb/Downloads/releng-svn/thermite-r688@136 5711f056-dcf5-4cc1-a278-612e92bcda7e

+++ user/gjb/scripts/release/ftp-stage.sh
@@ -178,7 +178,7 @@ create_dirs() {
 }
 
 create_dir_symlinks() {
-	if [ ! -z ${backpath} ]; then
+	if [ ! -z ${backpath} ] && [ ! -L ${_ftpdir}/${backpath}/${OSRELEASE} ]; then
 		echo "=== Creating backwards-compatible symlink:"
 		echo "==== ${backpath}/${OSRELEASE} -> ${path}/${OSRELEASE}"
 		ln -sf ${backpath}/${OSRELEASE} ${_ftpdir}/${backpath}/${OSRELEASE}
